+++ solvis/inversion_solution/fault_system_solution.py
@@ -1,3 +1,4 @@
+import io
 import logging
 import zipfile
 from pathlib import Path
@@ -19,11 +20,9 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
 
     _composite_rates: pd.DataFrame = ...
 
-    # def __init__(self, source_logic_tree):
-    #     self._source_logic_tree = source_logic_tree
-    #     self._solutions = {}
-
-    def set_props(self, composite_rates, aggregate_rates, ruptures, indices, fault_sections, fault_regime):
+    def set_props(
+        self, composite_rates, aggregate_rates, ruptures, indices, fault_sections, fault_regime, average_slips
+    ):
         # self._init_props()
         self._composite_rates = composite_rates
         self._aggregate_rates = aggregate_rates
@@ -31,16 +30,24 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
         self._fault_sections = fault_sections
         self._indices = indices
         self._fault_regime = fault_regime
+        self._average_slips = average_slips
+
+        # Now we need a rates table, structured correctly, with weights from the aggregate_rates
+        rates = self.aggregate_rates.drop(columns=['rate_max', 'rate_min', 'rate_count', 'fault_system']).rename(
+            columns={"rate_weighted_mean": "Annual Rate"}
+        )
+        self._rates = rates
 
     @staticmethod
-    def from_archive(instance_or_path: Union[Path, str, zipfile.ZipFile]) -> 'FaultSystemSolution':
+    def from_archive(instance_or_path: Union[Path, str, io.BytesIO]) -> 'FaultSystemSolution':
         new_solution = FaultSystemSolution()
 
         # TODO: sort out this weirdness
-        if isinstance(instance_or_path, zipfile.ZipFile):
-            assert 'ruptures/fast_indices.csv' in instance_or_path.namelist()
-            assert 'composite_rates.csv' in instance_or_path.namelist()
-            assert 'aggregate_rates.csv' in instance_or_path.namelist()
+        if isinstance(instance_or_path, io.BytesIO):
+            with zipfile.ZipFile(instance_or_path, 'r') as zf:
+                assert 'ruptures/fast_indices.csv' in zf.namelist()
+                assert 'composite_rates.csv' in zf.namelist()
+                assert 'aggregate_rates.csv' in zf.namelist()
             new_solution._archive = instance_or_path
         else:
             assert zipfile.Path(instance_or_path, at='ruptures/fast_indices.csv').exists()
@@ -56,17 +63,44 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
         cr = solution.composite_rates
         ar = solution.aggregate_rates
         ri = solution.indices
+        avs = solution.average_slips
+
         ruptures = rr[rr["Rupture Index"].isin(rupture_ids)].copy()
         composite_rates = cr[cr["Rupture Index"].isin(rupture_ids)].copy()
         aggregate_rates = ar[ar["Rupture Index"].isin(rupture_ids)].copy()
         indices = ri[ri["Rupture Index"].isin(rupture_ids)].copy()
+        average_slips = avs[avs["Rupture Index"].isin(rupture_ids)].copy()
 
         # all other solution properties are derived from those above
         ns = FaultSystemSolution()
         ns.set_props(
-            composite_rates, aggregate_rates, ruptures, indices, solution.fault_sections.copy(), solution.fault_regime
+            composite_rates,
+            aggregate_rates,
+            ruptures,
+            indices,
+            solution.fault_sections.copy(),
+            solution.fault_regime,
+            average_slips,
         )
-        ns._archive_path = solution._archive_path
+        # ns._archive_path = None
+        ns.enable_fast_indices()
+        # copy the original archive, if it exists
+        # TODO: does the archive needs filtering applied?? see to_archive()
+        if solution._archive:
+            new_archive = io.BytesIO()
+            with zipfile.ZipFile(new_archive, 'w') as new_zip:
+                # write the core files
+                with zipfile.ZipFile(solution._archive, 'r') as zf:
+                    for item in zf.filelist:
+                        if item.filename in solution.DATAFRAMES:
+                            continue
+                        if item.filename in solution.OPENSHA_ONLY:  # drop bulky, opensha-only artefacts
+                            continue
+                        new_zip.writestr(item, zf.read(item.filename))
+                # write the modifies tables
+                ns._write_dataframes(new_zip, reindex=False)  # retain original rupture ids and structure
+            ns._archive = new_archive
+            ns._archive.seek(0)
         return ns
 
     @staticmethod
@@ -74,45 +108,20 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
         # build a new composite solution, taking solution template properties, and composite_rates_df
         ns = FaultSystemSolution()
 
-        # # Remove all ruptures with no rates
-        # print('COMP 0', composite_rates_df.shape)
-        # print(composite_rates_df)
-        # print(composite_rates_df.info())
-        # composite_rates_df.to_json('diag_composite_rates_df_0.json')
-
         composite_rates_df = composite_rates_df[composite_rates_df["Annual Rate"] > 0]
-
-        # TODO CBC/CDC -use the weight column on composite_rates_df to do weighted mean etc
-        weighted_rate = pd.Series(composite_rates_df['Annual Rate'] * composite_rates_df['weight'], dtype="float32")
-        composite_rates_df.insert(0, 'weighted_rate', weighted_rate)
-
-        # print('COMP 1', composite_rates_df.shape)
-        # print(composite_rates_df)
-        # print(composite_rates_df.info())
-        # composite_rates_df.to_json('diag_composite_rates_df_1.json')
+        composite_rates_df.insert(
+            0,
+            'weighted_rate',
+            pd.Series(composite_rates_df['Annual Rate'] * composite_rates_df['weight'], dtype="float32"),
+        )
 
         aggregate_rates_df = composite_rates_df.pivot_table(
-            # values=['weighted_rate', "Annual Rate"],
-            index=[
-                'fault_system',
-                'Rupture Index',
-            ],
+            index=['fault_system', 'Rupture Index'],
             aggfunc={"Annual Rate": [np.min, np.max, 'count'], "weighted_rate": np.sum},
         )
 
-        # #aggregate_rates_df.to_json('diag_aggregate_rates.json')
-        # print('AGG 0', aggregate_rates_df.shape)
-        # print(aggregate_rates_df)
-        # print(aggregate_rates_df.info())
-
         # drop the top index level
         aggregate_rates_df.columns = aggregate_rates_df.columns.get_level_values(1)
-
-        # print('AGG', aggregate_rates_df.shape)
-        # print(aggregate_rates_df)
-        # print(aggregate_rates_df.info())
-
-        # rename the aggregate columns
         aggregate_rates_df = aggregate_rates_df.reset_index().rename(
             columns={
                 "amax": "rate_max",
@@ -121,16 +130,7 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
                 "sum": "rate_weighted_mean",
             }
         )
-
-        # print()
-        # print('AGG 1', aggregate_rates_df.shape)
-        # print(aggregate_rates_df)
-        # print(aggregate_rates_df.info())
-
         composite_rates_df = composite_rates_df.drop(columns="weighted_rate")
-        # # debugggg
-        # composite_rates_df.to_json('diag_composite_rates_df.json')
-        # aggregate_rates_df.to_json('diag_aggregate_rates.json')
 
         ns.set_props(
             composite_rates_df,
@@ -139,8 +139,8 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
             solution.indices.copy(),
             solution.fault_sections.copy(),
             solution.fault_regime,
+            solution.average_slips.copy(),
         )
-
         return ns
 
     @staticmethod
@@ -163,9 +163,6 @@ class FaultSystemSolution(FaultSystemSolutionFile, InversionSolutionOperations):
             composite_rates_df = pd.concat([composite_rates_df, solution_df], ignore_index=True)
 
             # print('dims', composite_rates_df.shape, solution_df.shape)
-
-        # composite_rates_df.solution_id = composite_rates_df.solution_id.astype('category')
-        # composite_rates_df.fault_system = composite_rates_df.fault_system.astype('category')
         return FaultSystemSolution.new_solution(solution=branch_solution, composite_rates_df=composite_rates_df)
 
     @property