+++ solvis/inversion_solution/fault_system_solution_file.py
@@ -6,7 +6,7 @@ import geopandas as gpd
 import numpy as np
 import pandas as pd
 
-from .inversion_solution_file import InversionSolutionFile, data_to_zip_direct, reindex_dataframe
+from .inversion_solution_file import InversionSolutionFile, data_to_zip_direct
 
 log = logging.getLogger(__name__)
 
@@ -38,44 +38,25 @@ class FaultSystemSolutionFile(InversionSolutionFile):
 
     FAST_INDICES_PATH = 'ruptures/fast_indices.csv'  # store indices in standard table format for much faster I/O
 
-    DATAFRAMES = [
-        # InversionSolutionFile.RATES_PATH,
-        InversionSolutionFile.RUPTS_PATH,
-        InversionSolutionFile.INDICES_PATH,
+    # OPENSHA_ONLY
+    # OPENSHA_LTB_PATH = 'ruptures/logic_tree_branch.json'
+    OPENSHA_SECT_POLYS_PATH = 'ruptures/sect_polygons.geojson'
+    OPENSHA_GRID_REGION_PATH = 'ruptures/grid_region.geojson'
+    OPENSHA_ONLY = [OPENSHA_SECT_POLYS_PATH, OPENSHA_GRID_REGION_PATH]
+
+    DATAFRAMES = InversionSolutionFile.DATAFRAMES + [
         COMPOSITE_RATES_PATH,
         AGGREGATE_RATES_PATH,
         FAST_INDICES_PATH,
     ]
 
     def _write_dataframes(self, zip_archive: zipfile.ZipFile, reindex: bool = False):
+        data_to_zip_direct(zip_archive, self._composite_rates.to_csv(index=reindex), self.COMPOSITE_RATES_PATH)
+        data_to_zip_direct(zip_archive, self._aggregate_rates.to_csv(index=reindex), self.AGGREGATE_RATES_PATH)
+        if self._fast_indices is not None:
+            data_to_zip_direct(zip_archive, self._fast_indices.to_csv(index=reindex), self.FAST_INDICES_PATH)
 
-        log.info("%s write_dataframes with fast_indices: %s" % (type(self), self._fast_indices is not None))
-
-        if reindex:
-            data_to_zip_direct(
-                zip_archive, reindex_dataframe(self._composite_rates).to_csv(index=True), self.COMPOSITE_RATES_PATH
-            )
-            data_to_zip_direct(
-                zip_archive, reindex_dataframe(self._aggregate_rates).to_csv(index=True), self.AGGREGATE_RATES_PATH
-            )
-            # data_to_zip_direct(zip_archive, reindex_dataframe(self._rates).to_csv(), self.RATES_PATH)
-            data_to_zip_direct(zip_archive, reindex_dataframe(self._ruptures).to_csv(index=False), self.RUPTS_PATH)
-            if self._fast_indices is not None:
-                data_to_zip_direct(
-                    zip_archive, reindex_dataframe(self._fast_indices).to_csv(index=False), self.FAST_INDICES_PATH
-                )
-            else:
-                data_to_zip_direct(zip_archive, reindex_dataframe(self._indices).to_csv(index=False), self.INDICES_PATH)
-
-        else:
-            data_to_zip_direct(zip_archive, self._composite_rates.to_csv(index=False), self.COMPOSITE_RATES_PATH)
-            # data_to_zip_direct(zip_archive, self._rates.to_csv(index=False), self.RATES_PATH)
-            data_to_zip_direct(zip_archive, self._aggregate_rates.to_csv(index=False), self.AGGREGATE_RATES_PATH)
-            data_to_zip_direct(zip_archive, self._ruptures.to_csv(index=False), self.RUPTS_PATH)
-            if self._fast_indices is not None:
-                data_to_zip_direct(zip_archive, self._fast_indices.to_csv(index=False), self.FAST_INDICES_PATH)
-            else:
-                data_to_zip_direct(zip_archive, self._indices.to_csv(index=False), self.INDICES_PATH)
+        super()._write_dataframes(zip_archive, reindex)
 
     def to_archive(self, archive_path, base_archive_path, compat=False):
         """
@@ -106,10 +87,6 @@ class FaultSystemSolutionFile(InversionSolutionFile):
         dtypes["fault_system"] = pd.CategoricalDtype()
         df = self._dataframe_from_csv(self._aggregate_rates, self.AGGREGATE_RATES_PATH, dtypes)
         return df.set_index(["fault_system", "Rupture Index"], drop=False)
-
-    # @property
-    # def rupture_rates(self):
-    #     raise NotImplementedError("Use aggregate_rates instead")
 
     @property
     def fast_indices(self) -> gpd.GeoDataFrame: